Remove binding attribute inheritence to avoid binding rule confusion
